// app/services/types/home.ts
export interface PlaceCODOrderRequest {
  facility_id: number;
  user_id: number;
  mobile: string;
  user_address_id: number;
  sub_total: number;
  discount: number;
  tax_percentage: number;
  total_amount: number;
  payment_method: string;
  payment_id: string;
  transaction_id: number;
  driver_notes: string;
  facility_notes: string;
  schedule: {
    pickup: {
      date: string;
      slot: string;
      time: string;
      collect_from_id: number;
    };
    drop: {
      date: string;
      slot: string;
      time: string;
      collect_from_id: number;
    };
  };
  items: {
    item_id: number;
    amount: number;
    quantity: number;
    service_id: number;
    category_id: number;
  }[];
}
// ...
export interface PlaceOnlineOrderRequest {
  amount: string;
}